srpic engine modularized
